+++ src/app/shared/data/LandingPageData.ts
