+++ src/app/shared/data/StaticsComponentData.ts
