+++ src/app/shared/data/galleryData.ts
