+++ src/app/shared/data/insightSectionData.ts
