+++ src/app/shared/types/LandingPageTypes.ts
