+++ src/app/shared/types/StaticsComponentDataTypes.ts
