+++ src/app/shared/types/classesTypes.ts
