+++ src/app/shared/types/galleryDataTypes.ts
