+++ src/app/shared/types/insightSectionDataTypes.ts
